+++ R/aoa.R
@@ -2,7 +2,7 @@
 #' @description
 #' This function estimates the Dissimilarity Index (DI) and the derived
 #' Area of Applicability (AOA) of spatial prediction models by
-#' considering the distance of new data (i.e. a Raster Stack of spatial predictors
+#' considering the distance of new data (i.e. a SpatRaster of spatial predictors
 #' used in the models) in the predictor variable space to the data used for model
 #' training. Predictors can be weighted based on the internal
 #' variable importance of the machine learning algorithm used for model training.
@@ -64,7 +64,7 @@
 #' set.seed(100)
 #' pts <- pts[1:30,]
 #' studyArea <- rast(system.file("extdata","predictors_2012-03-25.grd",package="CAST"))[[1:8]]
-#' trainDat <- extract(studyArea,pts,df=TRUE)
+#' trainDat <- extract(studyArea,pts,na.rm=FALSE)
 #' trainDat <- merge(trainDat,pts,by.x="ID",by.y="ID")
 #'
 #' # visualize data spatially:
@@ -146,7 +146,6 @@ aoa <- function(newdata,
 
   # handling of different raster formats
   as_stars <- FALSE
-  as_raster <- FALSE
   leading_digit <- any(grepl("^{1}[0-9]",names(newdata)))
 
   if (inherits(newdata, "stars")) {
@@ -156,11 +155,10 @@ aoa <- function(newdata,
     as_stars <- TRUE
   }
   if (inherits(newdata, "Raster")) {
-    if (!requireNamespace("raster", quietly = TRUE))
-      stop("package raster required: install that first")
+   # if (!requireNamespace("raster", quietly = TRUE))
+  #    stop("package raster required: install that first")
     message("Raster will soon not longer be supported. Use terra or stars instead")
     newdata <- methods::as(newdata, "SpatRaster")
-    as_raster <- TRUE
   }
 
 
@@ -197,7 +195,6 @@ aoa <- function(newdata,
   if (inherits(newdata, "SpatRaster")){
     if (any(is.factor(newdata))){
       newdata[[which(is.factor(newdata))]] <- as.numeric(newdata[[which(is.factor(newdata))]])
-      #newdata[[which(is.factor(newdata))]] <- raster::deratify(newdata[[which(is.factor(newdata))]],complete = TRUE)
     }
     newdata <- terra::as.data.frame(newdata,na.rm=FALSE)
   }
@@ -282,10 +279,6 @@ aoa <- function(newdata,
     if (as_stars){
       out <- stars::st_as_stars(out)
       AOA <- stars::st_as_stars(AOA)
-    }
-    if(as_raster){
-      out <- methods::as(out, "Raster")
-      AOA <- methods::as(AOA, "Raster")
     }
 
   }else{