+++ R/plot.R
@@ -47,16 +47,18 @@ plot.aoa = function(x, samplesize = 1000, ...){
 
 
   if(inherits(x$AOA, "RasterLayer")){
-    targetDI = raster::sampleRegular(x$DI, size = samplesize)
-    targetDI = data.frame(DI = as.numeric(targetDI),
+    targetDI = terra::spatSample(methods::as(x$DI, "SpatRaster"),
+                                 size = samplesize, method="regular")
+    targetDI = data.frame(DI = as.numeric(targetDI[,1]),
                           what = "predictionDI")
   }else if(inherits(x$AOA, "stars")){
-    targetDI = raster::sampleRegular(methods::as(x$DI, "Raster"), size = samplesize)
-    targetDI = data.frame(DI = as.numeric(targetDI),
+    targetDI = terra::spatSample(methods::as(x$DI, "SpatRaster"),
+                                 size = samplesize,method="regular")
+    targetDI = data.frame(DI = as.numeric(targetDI[,1]),
                           what = "predictionDI")
   }else if(inherits(x$AOA, "SpatRaster")){
-    targetDI = raster::sampleRegular(methods::as(x$DI, "Raster"), size = samplesize)
-    targetDI = data.frame(DI = as.numeric(targetDI),
+    targetDI = terra::spatSample(x$DI, size = samplesize,method="regular")
+    targetDI = data.frame(DI = as.numeric(targetDI[,1]),
                           what = "predictionDI")
   }else{
     targetDI = data.frame(DI = sample(x$DI, size = samplesize),